% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{print.default}
\alias{print.default}
\title{Default print method.}
\usage{
\method{print}{default}(x, ...)
}
\arguments{
\item{x}{print object}

\item{...}{other parameters}
}
\description{
Default print method.
}
