% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{print.SNACSList}
\alias{print.SNACSList}
\title{Prints SNACSList object.}
\usage{
\method{print}{SNACSList}(x, ...)
}
\arguments{
\item{x}{SNACSList}

\item{...}{ignored}
}
\value{
A SNACSList object
}
\description{
Prints SNACSList object.
}
