% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{SNACSList}
\alias{SNACSList}
\title{Creates a SNACSList object.}
\usage{
SNACSList(x, y)
}
\arguments{
\item{x}{Matrix having mutation status as 0s and 1s}

\item{y}{Matrix having hash values as numeric data}
}
\value{
A SNACSList object
}
\description{
SNACS stores data in a simple list-based data object called a SNACSList.
}
\examples{
z <- SNACSList(x = 1, y = 2)
print(z)

}
